OT231-25 : Edit and Move Security Authentication Add UserResponse
