--- src/main/java/com/alkemy/ong/infrastructure/config/spring/models/AuthenticationResponse.java
@@ -1,14 +1,0 @@
-package com.alkemy.ong.infrastructure.config.spring.models;
-
-public class AuthenticationResponse {
-
-  private final String jwt;
-
-  public AuthenticationResponse(String jwt) {
-    this.jwt = jwt;
-  }
-
-  public String getJwt() {
-    return jwt;
-  }
-}
