+++ src/main/java/com/alkemy/ong/infrastructure/config/spring/security/SecurityConfig.java
@@ -57,7 +57,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
   @Override
   protected void configure(HttpSecurity http) throws Exception {
 
-    http.authorizeRequests()
+    http.csrf().disable().authorizeRequests()
             .antMatchers("/auth/login")
             .permitAll()
             .anyRequest().authenticated();
