+++ src/main/java/com/alkemy/ong/infrastructure/rest/response/AuthenticationResponse.java
@@ -1,0 +1,10 @@
+package com.alkemy.ong.infrastructure.rest.response;
+
+import lombok.Data;
+
+@Data
+public class AuthenticationResponse {
+
+  private final String jwt;
+
+}
